% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goldenSection.R
\name{golden_section}
\alias{golden_section}
\title{Golden Section}
\usage{
golden_section(f, interval, tol = 1e-08, max_iter = 1000)
}
\arguments{
\item{f}{the function to minimize/ maximize}

\item{interval}{the interval over which the function is to be evaluated}

\item{tol}{the toloerance}

\item{max_iter}{maximum number of iterations}
}
\value{
a list of useful statistics
}
\description{
Golden Section
}
\examples{
f <- function(x) (x - 2)^2
result <- golden_section(f, c(0, 5))
print(result)
}
